+++ lib/assets/javascripts/cartodb3/editor/style/style-form/style-aggregation-form/style-aggregation-properties-form-model.js
@@ -13,7 +13,7 @@ module.exports = cdb.core.Model.extend({
   },
 
   _initBinds: function () {
-    this.bind('change', this._onChange, this);
+    this.bind('change', _.debounce(this._onChange.bind(this), 150), this);
   },
 
   _onChange: function () {
