+++ lib/assets/javascripts/cartodb3/editor/style/style-form/style-form-view.js
@@ -1,6 +1,6 @@
 var cdb = require('cartodb.js');
 var _ = require('underscore');
-var StylePropertiesFormView = require('./properties-style-form/style-properties-form-view');
+var StylePropertiesFormView = require('./style-properties-form/style-properties-form-view');
 var StyleAggregationFormView = require('./style-aggregation-form/style-aggregation-form-view');
 var styleFormNotReadyTemplate = require('./style-form-not-ready.tpl');
