+++ lib/assets/javascripts/cartodb3/editor/style/style-form/style-properties-form/style-properties-form.tpl
