+++ lib/assets/javascripts/cartodb3/editor/style/style-form/style-properties-form/style-shape-properties-form-view.js
